Refactor filterSelectOptions to accept also function
The filterSelectOptions property in MRT_ColumnDef has been refactored to accept either an array of DropdownOption or a function that takes an array of keys and returns an array of DropdownOption. This change allows for more flexibility in providing select options for filtering. The code in column.utils.ts has been updated to handle both cases.